add option to output list of files in ecco_download module
